requirements.txt
Flask
requests
beautifulsoup4
google-generativeai
pymongo
python-dotenv
scikit-learn
tenacity
pandas
openpyxl
aiohttp